Omfattande ändringar för användning av doubles.
AnkomstEvent.java: Kund får nu en referens till MarketState här. Se
kund.java för mer information.

Event.java: time är nu en double, och dess returmetod har bytt returtyp.

EventQueue.java: ArrayList sortedTime är nu en array av doubles. Den
interna variabeln time har även den förändrats till en double.

Kund.java: Konstruktorn behåller nu en referens till MarketState.
MarketState har metoder som genererar respektive tider från det frö och
intervall som skickats in. Alla tider lagras nu som doubles.

MarketEvent.java: Bytte jämförelsen på rad 48 från mindre eller lika med
till lika med. Om antalet lediga kassor är mindre än noll har något gått
fel och programmet bör krascha eller flippa ut på ett sätt som
reflekterar det.

MarketState.java: Statistikvariabler har bytt typ från int till double.

MarketView.java: Använder sig nu av ett objekt som avrundar tal med
decimaler till önskat format. Några buggfixar då event inte skrevs ut på
rätt sätt, och results inte alls överhuvudtaget.

PlockEvent.java: Lite mer whitespace i dess toString() för att
tabuleringen ska bli rätt.

StartEvent.java: La till ett anrop till IncomingEvent så att MarketView
skriver ut händelsen.

State.java: globalTime och timeMax är nu doubles.

StopEvent: Tog bort main-metoden. Gör nu anrop till incomingEvent samt
kommenterade bort spårutskriften (detta ska ske i MarketView).
Parametern i konstruktorn är nu en double.

StängerEvent.java: Tog bort main-metoden och bytte typ på parametern i
konstruktorn till double.

BetalaEvent.java: Interna variabeln queueTimer är nu en double, och
metoden timeChange tar nu en double som parameter.

+++ lab5/src/lab5/classtemplates/event/Event.java
@@ -18,12 +18,12 @@ public abstract class Event extends Observable{
 	private int executionTime;
 	public EventQueue eventQueue;
 	
-	public int time;
+	public double time;
 	/**
 	 * Förändrar hur lång tid som återstår innan eventet ska hända.
 	 * @param elapsedTime Mängden tid som förflutit.
 	 */
-	public void timeChange (int elapsedTime){
+	public void timeChange (double elapsedTime){
 		time -= elapsedTime;
 	}
 	/*
@@ -38,7 +38,7 @@ public abstract class Event extends Observable{
 	 * Hämtar tiden som återstår innan eventet händer.
 	 * @return Tiden som återstår.
 	 */
-	public int time() {
+	public double time() {
 		
 		return this.time;
 	}
